Remove old constraint JSON parsers
